src/streamlit_app/requirements.txt
numpy
pandas
plotly
streamlit
streamlit_plotly_events
streamlit
pandas
sqlalchemy
psycopg2-binary
plotly
openpyxl
python-dotenv